stock_api/requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.0.8
django-cors-headers==3.4.0
django-crontab==0.7.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.2
matplotlib==3.2.2
multitasking==0.0.9
mysqlclient==2.0.1
nltk==3.5
numpy==1.19.0
pandas==1.0.5
pandas-datareader==0.8.1
pinkfish==0.5.1
pkg-resources==0.0.0
protobuf==3.12.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow==2.0.0b1
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Theano==1.0.4
threadpoolctl==2.1.0
tqdm==4.47.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
yfinance==0.1.54
zipp==3.1.0
